Moved HttpAuth from plugins into auth
