+++ src/core/components/auth/index.js
@@ -6,7 +6,9 @@ export { default as AuthFormRow } from "./auth-form-row"
 export { default as AuthError } from "./auth-error"
 export { default as ApiKeyAuth } from "./auth-api-key"
 export { default as BasicAuth } from "./auth-basic"
-export { default as AuthItem } from "./auth-item"
+export { default as AuthHttp } from "./auth-http"
+export { default as AuthHttpBasic } from "./auth-http-basic"
+export { default as AuthHttpBearer } from "./auth-http-basic"
 export { default as Oauth2 } from "./auth-oauth2"
 export { default as AuthOauth2Info } from "./auth-oauth2-info"
 export { default as Oauth2Form } from "./auth-oauth2-form"
