+++ src/core/plugins/oas3/wrap-components/index.js
@@ -1,5 +1,4 @@
 import Markdown from "./markdown"
-import AuthItem from "./auth-item"
 import parameters from "./parameters"
 import VersionStamp from "./version-stamp"
 import OnlineValidatorBadge from "./online-validator-badge"
@@ -8,7 +7,6 @@ import JsonSchema_string from "./json-schema-string"
 
 export default {
   Markdown,
-  AuthItem,
   parameters,
   JsonSchema_string,
   VersionStamp,
