+++ src/core/presets/base.js
@@ -25,9 +25,11 @@ import {
   AuthFormRow,
   AuthError,
   AuthBtnGroup,
-  AuthItem,
   ApiKeyAuth,
   BasicAuth,
+  AuthHttp,
+  AuthHttpBasic,
+  AuthHttpBearer,
   Oauth2,
   AuthOauth2Info,
   Oauth2Form,
@@ -107,7 +109,6 @@ export default function() {
       AuthorizeBtnContainer,
       authorizeOperationBtn: AuthorizeOperationBtn,
       auths: Auths,
-      AuthItem: AuthItem,
       AuthBtnGroup,
       AuthFormRow,
       AuthOauth2Info,
@@ -117,6 +118,9 @@ export default function() {
       oauth2: Oauth2,
       apiKeyAuth: ApiKeyAuth,
       basicAuth: BasicAuth,
+      authHttp: AuthHttp,
+      authHttpBasic: AuthHttpBasic,
+      authHttpBearer: AuthHttpBearer,
       clear: Clear,
       liveResponse: LiveResponse,
       info: Info,
